+++ Arbiter.App/Models/GameClientWindow.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Arbiter.Interop.Window;
 
 namespace Arbiter.App.Models;
 
@@ -9,4 +10,15 @@ public class GameClientWindow
     public string WindowClassName { get; set; } = string.Empty;
     public string WindowTitle { get; set; } = string.Empty;
     public string? CharacterName { get; set; }
+
+    public void SetWindowTitle(string title)
+    {
+        NativeWindowHelper.SetWindowTitle(WindowHandle, title);
+        WindowTitle = title;
+    }
+
+    public void BringToFront()
+    {
+        NativeWindowHelper.BringToFront(WindowHandle);
+    }
 }